requirements.txt
python-docx # to check Word files
pymupdf     # to check PDF files
